+++ package.json
@@ -36,6 +36,7 @@
     "babel-loader": "^8.2.2",
     "lodash": "^4.17.21",
     "postcss-selector-parser": "^6.0.6",
-    "style-loader": "^2.0.0"
+    "style-loader": "^2.0.0",
+    "uuidv4": "^6.2.10"
   }
 }
