+++ src/index.js
@@ -1,61 +1,17 @@
 import './style.css';
 import displayData from './print';
-import { set } from 'lodash';
-
+import { Project, Todo, clearField, addOption, toggleDone, deleteTodo, clearFieldPro } from './logic.js';
 window.onload = function () {
 	const title = document.getElementById('title');
 	const description = document.getElementById('description');
 	const duedate = document.getElementById('duedate');
 	const addProject = document.getElementById('addProject');
-	const proTitle = document.getElementById('project-input');
+
 	let todoItems = [];
 	const projectName = [];
 	let stored_todo = JSON.parse(localStorage.getItem('todos'));
+
 	const form = document.querySelector('form');
-	const Project = function (title) {
-		this.title = title;
-	};
-
-	const Todo = function (title, description, duedate, project = 'defualt') {
-		this.title = title;
-		this.description = description;
-		this.duedate = duedate;
-		this.checked = false;
-		this.project = project;
-		this.id = Date.now();
-	};
-
-	function clearField() {
-		title.value = '';
-		description.value = '';
-		duedate.value = '';
-	}
-	function clearFieldPro() {
-		proTitle.value = '';
-	}
-	const addOption = function (pro) {
-		const selectOption = document.getElementById('project');
-
-		const opt = document.createElement('option');
-		opt.value = pro.title;
-		opt.innerHTML = pro.title;
-		selectOption.append(opt);
-	};
-
-	const toggleDone = function (key) {
-		const index = todoItems.findIndex((item) => item.id === Number(key));
-		todoItems[index].checked = !todoItems[index].checked;
-		displayData(todoItems[index]);
-	};
-	const deleteTodo = function (key) {
-		const index = todoItems.findIndex((item) => item.id === Number(key));
-		const todo = {
-			deleted: true,
-			...todoItems[index],
-		};
-		todoItems = todoItems.filter((item) => item.id !== Number(key));
-		displayData(todo);
-	};
 
 	form.addEventListener('submit', (event) => {
 		event.preventDefault();
@@ -69,13 +25,13 @@ window.onload = function () {
 		} else {
 			const newTodo = new Todo(title, description, duedate, project);
 			todoItems.push(newTodo);
+			setItem('todos', todoItems);
+			getItemFromLocalStorage('todos');
 
-			displayData(newTodo);
-
-			console.log(newTodo);
 			clearField();
 		}
-		setItem('todos', todoItems);
+
+
 
 	});
 
@@ -93,7 +49,7 @@ window.onload = function () {
 			clearFieldPro();
 		}
 		setItem('pros', projectName);
-		getItemFromLocalStorage('pros');
+
 
 	});
 
@@ -110,21 +66,7 @@ window.onload = function () {
 		}
 	});
 
-	//trying for accessing data from local storage
-	const getItemFromLocalStorage = function (names) {
-		let project_name = JSON.parse(localStorage.getItem(names));
-		//	let todos_item = JSON.parse(localStorage.getItem('todos'));
-		const project_list = document.createElement('li');
-		const lists = document.getElementById('content');
-		for (let i = 0; i < project_name.length; i++) {
-			project_list.innerHTML += project_name[i].title;
-			lists.appendChild(project_list);
-			console.log(project_name);
-		}
-	}
-	const setItem = function (name, arr) {
-		localStorage.setItem(name, JSON.stringify(arr));
 
-	}
-	displayData(stored_todo);
+
+
 };
